requirements.txt
fastapi
uvicorn
pillow
paddleocr
numpy
ultralytics
transformers
retinaface
python-multipart